word_floating_agent/requirements.txt
PySide6
langchain-openai
python-dotenv
pynput
darkdetect
pywin32; sys_platform == 'win32'
